+++ src/modules/questionnaires/questionnaire_log.ts
@@ -1,0 +1,1 @@
+export namespace QuestionnaireLog {}
