+++ src/modules/questionnaires/questionnaire_repository.ts
@@ -1,0 +1,12 @@
+import database from '../../config/database'
+import { QuestionnaireEntity } from './questionnaire_entity'
+
+export class QuestionnaireRepository {
+  private Villages = () => database<QuestionnaireEntity.Village>('villages')
+
+  public questionnaire = (id: string, properties: string) =>
+    this.Villages().where('id', id).update({
+      properties,
+      updated_at: new Date(),
+    })
+}
