+++ src/modules/questionnaires/questionnaire_response.ts
@@ -1,0 +1,1 @@
+export class QuestionnaireResponse {}
